+++ src/main/java/com/sr182022/travelagencystar/controller/TravelController.java
@@ -1,10 +1,13 @@
 package com.sr182022.travelagencystar.controller;
 
 import com.sr182022.travelagencystar.model.Travel;
+import com.sr182022.travelagencystar.model.User;
 import com.sr182022.travelagencystar.service.IAccommodationUnitService;
 import com.sr182022.travelagencystar.service.IDestinationService;
 import com.sr182022.travelagencystar.service.ITravelService;
 import com.sr182022.travelagencystar.service.IVehicleService;
+import com.sr182022.travelagencystar.utils.CheckRoleUtil;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -26,7 +29,11 @@ public class TravelController {
     }
 
     @GetMapping("/travel")
-    public String travelDetailsPage(@RequestParam int id, Model model) {
+    public String travelDetailsPage(HttpSession session, @RequestParam int id, Model model) {
+        if(!CheckRoleUtil.RoleOrganizer(session)) {
+            return "redirect:/permission-error";
+        }
+
         Travel travel = travelService.findOne(id);
         /*  if(travel == null) {
             return "redirect:/errorPages/routeErrorPage.html";
@@ -36,17 +43,30 @@ public class TravelController {
     }
 
     @PostMapping("/dashboard/travels/addNewTravel")
-    public String addNewDestination(@ModelAttribute Travel newTravel,
+    public String addNewDestination(HttpSession session, @ModelAttribute Travel newTravel,
                                     @RequestParam int destinationId,
                                     @RequestParam int accommodationUnitId,
                                     @RequestParam int vehicleId) {
+        if(!CheckRoleUtil.RoleOrganizer(session)) {
+            return "redirect:/permission-error";
+        }
+        if(newTravel.getStartDate() == null || newTravel.getEndDate() == null) {
+            // handle here
+        }
         newTravel.setNumberOfNights(travelService.setNumberOfNights(newTravel.getStartDate(), newTravel.getEndDate()));
         travelService.save(newTravel, destinationId, accommodationUnitId, vehicleId);
         return "redirect:/dashboard/travels";
     }
 
     @GetMapping("/dashboard/travels/editTravel")
-    public String editTravel(@RequestParam int travelId, Model model) {
+    public String editTravel(HttpSession session, @RequestParam int travelId, Model model) {
+        if(!CheckRoleUtil.RoleOrganizer(session)) {
+            return "redirect:/permission-error";
+        }
+        Travel travel = travelService.findOne(travelId);
+        if(travel == null) {
+            return "redirect:/route-error";
+        }
         model.addAttribute("travel", travelService.findOne(travelId));
         model.addAttribute("destinationsForSelectMenu", destinationService.findAll());
         model.addAttribute("vehiclesForSelectMenu", vehicleService.findAll());
@@ -56,14 +76,20 @@ public class TravelController {
     }
 
     @PostMapping("/dashboard/travels/editTravelPost")
-    public String editTravelPost(@ModelAttribute Travel editTravel, int destinationId, int accommodationUnitId, int vehicleId) {
+    public String editTravelPost(HttpSession session, @ModelAttribute Travel editTravel, int destinationId, int accommodationUnitId, int vehicleId) {
+        if(!CheckRoleUtil.RoleOrganizer(session)) {
+            return "redirect:/permission-error";
+        }
         editTravel.setNumberOfNights(travelService.setNumberOfNights(editTravel.getStartDate(), editTravel.getEndDate()));
         travelService.update(editTravel, destinationId, accommodationUnitId, vehicleId);
         return "redirect:/dashboard/travels";
     }
 
     @PostMapping("/dashboard/travels/deleteTravel")
-    public String deleteTravel(@RequestParam int travelId) {
+    public String deleteTravel(HttpSession session, @RequestParam int travelId) {
+        if(!CheckRoleUtil.RoleOrganizer(session)) {
+            return "redirect:/permission-error";
+        }
         travelService.delete(travelId);
         return "redirect:/dashboard/travels";
     }
